% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addFilter.R
\name{addFilter}
\alias{addFilter}
\title{Add a file type to file extensions}
\usage{
addFilter(rowname = "csv", desc = "csv files (*.csv)", filter = "*.csv",
  Filters = NULL)
}
\arguments{
\item{rowname}{- row name to add in filters matrix}

\item{desc}{- brief description of the file types}

\item{filter}{- extension filter ("*.ext")}

\item{Filters}{- matrix of filters to add to}
}
\value{
augmented Filters matrix
}
\description{
Add a filter to the file extension matrix
(if it's not there already) used in file dialogs.
}

