% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TextFunctions.R
\name{addQuotes}
\alias{addQuotes}
\title{Add quotes to a character string vector.}
\usage{
addQuotes(str, qt = "'")
}
\arguments{
\item{str}{- character vector to add quotes to}

\item{qt}{- quote character}
}
\value{
character vector w/ elements in quotes
}
\description{

}

