% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcInitSizeDist.R
\name{calcEquilSizeDist.11}
\alias{calcEquilSizeDist.11}
\title{Calculate the equilibrium size distribution for a model with
1 shell condition and 1 maturity state.}
\usage{
calcEquilSizeDist.11(r, A, S)
}
\arguments{
\item{r}{- vector of equilibrium recruits-at-size}

\item{A}{- size transition matrix}

\item{S}{- vector or diagonal matrix of survival by size class}
}
\value{
list with equilibrium numbers-at-size vectors for new shell and old shell crab
}
\description{
Function to calculate the equilibrium size distribution for a model with
1 shell conditions and 1 maturity state.
}
\details{

}

