% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcFishingMortalities.R
\name{calcFishingMortalities}
\alias{calcFishingMortalities}
\title{Calculate size-specific fishery capture and retention rates.}
\usage{
calcFishingMortalities(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}
}
\value{
list with the following elements
sel_fyxmsz - size-specific selectivity
ret_fyxmsz - size-specific retention
F_fyxms - fully-selected fishing capture rate
FC_fyxmsz - size-specific fishing capture rate
FM_fyxmsz - size-specific fishing mortality rate
RM_fyxmsz - size-specific retention mortality rate
DM_fyxmsz - size-specific discard mortality rate
}
\description{
Function to calculate size-specific fishery capture and retention rates.
}
\details{

}

