% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcInitSizeDist.R
\name{calcInitSizeDist}
\alias{calcInitSizeDist}
\title{Calculate initial size distributions.}
\usage{
calcInitSizeDist(mc, mp, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}
}
\value{
n_xmsz: 4-d array of initial population abundance by sex/maturity/shell condition/size
}
\description{
Function to calculate initial size distributions.
}
\details{

}

