% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcMaturity.R
\name{calcMaturity}
\alias{calcMaturity}
\title{Calculate maturity from model parameters.}
\usage{
calcMaturity(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{model configuration list object}

\item{showPlot}{- flag (T/F) to plot results}
}
\value{
mat_yxz
}
\description{
Function to calculate maturity from model parameters.
}
\details{


Interpretation differs between king and Tanner crab models. For king crab, the
results represent sex-specific maturity ogives (by size). For Tanner crab, the results
represent the sex-specific probability that an immature crab will molt to maturity (by size).
}

