% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcModelProcesses.R
\name{calcModelProcesses}
\alias{calcModelProcesses}
\title{Calculate model processes}
\usage{
calcModelProcesses(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
mp - list object for model processes
}
\description{
Function to calculate model processes prior to running the model
}
\details{

}

