% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcNM.R
\name{calcNaturalMortality}
\alias{calcNaturalMortality}
\title{Calculate natural mortality.}
\usage{
calcNaturalMortality(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
array M_yxmsz
}
\description{

}

