% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcNatZ.R
\name{calcNatZ}
\alias{calcNatZ}
\title{Project population numbers at size forward through the model time interval}
\usage{
calcNatZ(mc, mp, iN_xmsz, R_yxz, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{iN_xmsz}{- initial numbers at size array}

\item{R_yxz}{- recruitment at size array}
}
\value{
N_yxmsz: 5-d array of population numbers by year/sex/maturity/shell condition/size
}
\description{
Function to project population numbers at size forward through the model time interval
}
\details{

}

