% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcNatZ.Surveys.R
\name{calcNatZ.Surveys}
\alias{calcNatZ.Surveys}
\title{Calculate survey numbers at size through the model time interval}
\usage{
calcNatZ.Surveys(mc, mp, N_yxmsz, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{N_yxmsz}{- initial numbers at size array}
}
\value{
N_vyxmsz: 6-d array of survey numbers by year/sex/maturity/shell condition/size
}
\description{
Function to calculate survey numbers at size through the model time interval
}
\details{

}

