% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcNM.R
\name{calcNM.gmacs}
\alias{calcNM.gmacs}
\title{Calculate natural mortality for gmacs models.}
\usage{
calcNM.gmacs(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
array M_yxmsz
}
\description{

}

