% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcNM.R
\name{calcNM.tcsam}
\alias{calcNM.tcsam}
\title{Calculate natural mortality for tcsam models.}
\usage{
calcNM.tcsam(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
array M_yxmsz
}
\description{

}

