% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcPrMolt.R
\name{calcPrMolt}
\alias{calcPrMolt}
\title{Calculate probability of molting at size for sex, maturity state, shell condition.}
\usage{
calcPrMolt(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}
}
\value{
prMolt_xmsz
}
\description{

}

