% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRatZ.init}
\alias{calcRatZ.init}
\title{Calculate initial proportions recruiting-at-size}
\usage{
calcRatZ.init(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
R_z: 1d array with initial proportions of crab recruiting by size
}
\description{

}

