% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRatZ}
\alias{calcRatZ}
\title{Calculate proportions recruiting-at-size}
\usage{
calcRatZ(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
R_yz: d array with annual proportions of crab recruiting by size
}
\description{

}

