% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcRecruitment.R
\name{calcRecruitment}
\alias{calcRecruitment}
\title{Calculate recruitment-at-size by year, sex}
\usage{
calcRecruitment(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
R_yxz: 3d array with numbers  of crab recruiting by year/sex/size
}
\description{

}
\details{
R_yxz has units of exp(lnR). If lnR is in log(millions), then R_yxz
has units of millions of individuals.
}

