% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcSurveyCatchabilities.R
\name{calcSurveyCatchabilities}
\alias{calcSurveyCatchabilities}
\title{Calculate size-specific survey catchabilities.}
\usage{
calcSurveyCatchabilities(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{showPlot}{- flag to show plots}
}
\value{
list with the following elements:
sel_vyxmsz - size-specific selectivity
Q_vyxms - fully-selected survey catchability
Q_vyxmsz - size-specific survey catchability
}
\description{
Function to calculate size-specific survey catchabilities.
}
\details{

}

