% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcWatZ.R
\name{calcWatZ}
\alias{calcWatZ}
\title{Calculate weight-at-size by year, sex}
\usage{
calcWatZ(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
W_yxmsz: 5d array with weight-at-size by year/sex/maturity state/shell condition
}
\description{

}

