% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcZTM.R
\name{calcZTM}
\alias{calcZTM}
\title{Calculate size transition matrices by year, sex}
\usage{
calcZTM(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
prZAM_yxmszz: 5d array with weight-at-size by year/sex/maturity state/shell condition
}
\description{

}

