% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkKeyword.R
\name{checkKeyword}
\alias{checkKeyword}
\title{Check string against keyword and stop if unequal.}
\usage{
checkKeyword(chk, key)
}
\arguments{
\item{chk}{- string to check}

\item{key}{- key word to check against}
}
\description{
Function to check string against a keyword and stop if unequal.
}
\details{

}

