% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dimArray.R
\name{dimArray}
\alias{dimArray}
\title{Get an array}
\usage{
dimArray(mc, str, val = 0)
}
\arguments{
\item{mc}{- model configurqtion object}

\item{str}{- string describing array indices}
}
\value{
array of appropriate dimensions filled with default value(s)
}
\description{

}

