% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/first_difference.R
\name{first_difference}
\alias{first_difference}
\title{Calculate the first difference of a vector.}
\usage{
first_difference(x)
}
\arguments{
\item{x}{- the vector to calculate the first difference for}
}
\value{
the vector of first differences
}
\description{
Function to calculate the first difference of a vector.
}
\details{

}

