% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeMultipleRuns.BBRKC.R
\name{makeMultipleRuns.BBRKC}
\alias{makeMultipleRuns.BBRKC}
\title{Make multiple simulation runs.}
\usage{
makeMultipleRuns.BBRKC(nSim = 1, seeds = NA, inpfn = NULL,
  outfn = "gmacs.input.dat", topDir = ".", subDirBase = "BBRKC",
  showPlot = FALSE, pdf = NULL, width = 8, height = 6)
}
\arguments{
\item{nSim}{- number of simulations to run}

\item{seeds}{- vector (or NA or NULL) of random number generator seeds}

\item{inpfn}{- input filename for rsim model configuration}

\item{outfn}{- filename for output file (input file to gmacs)}

\item{topDir}{- top level directory/folder for saving results}

\item{subDirBase}{- base name for sequentially-numbered sub-directories/folders}

\item{showPlot}{- flag (T/F) to show/print plots}

\item{pdf}{- name of pdf file in each subfolder to save plots to (if not NULL)}

\item{width}{- page width for pdf}

\item{height}{- page height for pdf}
}
\value{
vector of seed values used (or NULL, if specified as such).
}
\description{
Funtion to make multiple simulation runs, saving the Gmacs input file and simulation output from each in a separate folder.
}
\details{


Results from each model run is written to a separate (sequentially-numbered) subfolder
below that specified by topDir. If showPlot=TRUE, plots for a model run are saved to a pdf file
if 'pdf' is not NULL, otherwise plots are shown on the screen device.

If seeds is NA, then nSim seeds are drawn from a uniform distribution [1,100000000].
If seeds is NULL, then each seed is based on the system time.
}

