% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/midpoints.R
\name{midpoints}
\alias{midpoints}
\title{Calculate the midpoints of a vector.}
\usage{
midpoints(x)
}
\arguments{
\item{x}{- the vector to calculate the midpoints for}
}
\value{
the vector of midpoints
}
\description{
Function to calculate the midpoints of a vector.
}
\details{

}

