% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseMC.Dims.R
\name{parseMC.Dims}
\alias{parseMC.Dims}
\title{Parse model dimensions information from a character string vector.}
\usage{
parseMC.Dims(rz, i)
}
\arguments{
\item{rz}{- character string vector from which to extract model dimensions info}

\item{i}{- index into character string vector at which to start parsing}
}
\value{
a list with elements
dims : model dimensions list object
i    : final index + 1 into rz at end of parsing dims info
}
\description{

}

