% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseMC.Fisheries.R
\name{parseMC.Fisheries}
\alias{parseMC.Fisheries}
\title{Parse fisheries section of a model configuration file.}
\usage{
parseMC.Fisheries(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of fisheries section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
fisheries - list object with fisheries info
}
\description{
Function to parse the 'fisheries' section of a model configuration file.
}
\details{

}

