% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseMC.giData.R
\name{parseMC.giData}
\alias{parseMC.giData}
\title{Parse growth increment data section of a model configuration file.}
\usage{
parseMC.giData(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of surveys section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
giData - list object with giData info
}
\description{
Function to parse the growth increment data ('giData') section of a model configuration file.
}
\details{

}

