% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parseMC.Surveys.R
\name{parseMC.Surveys}
\alias{parseMC.Surveys}
\title{Parse surveys section of a model configuration file.}
\usage{
parseMC.Surveys(rsp, i, dims)
}
\arguments{
\item{rsp}{- parsed text list from model configuration file}

\item{i}{- index to start of surveys section in rsp}

\item{dims}{- model configuration dims list}
}
\value{
list with
i - index starting next section
surveys - list object with surveys info
}
\description{
Function to parse the 'surveys' section of a model configuration file.
}
\details{

}

