% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseNum.R
\name{parseNum}
\alias{parseNum}
\title{Parse numeric expressions from a character vector.}
\usage{
parseNum(str)
}
\arguments{
\item{str}{- character vector to parse}
}
\value{
parsed and evaluated numerical expressions as a vector
}
\description{

}

