% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SelectivityFunctions.R
\name{plogis}
\alias{plogis}
\title{Calculate the logistic function}
\usage{
plogis(z, mu, sd)
}
\arguments{
\item{z}{- vector of sizes at which to compute selectivities}

\item{mu}{- size at which selectivity  = 0.5 (logit-scale mean)}

\item{sd}{- standard deviation in selectivity (logit-scale standard deviation)}
}
\value{
vector with selectivity values at the elements of z
}
\description{
Function to calculate the logistic function
}
\details{

}

