% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Run the simulation model}
\usage{
runModel(mc, mp, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}

\item{mp}{- model processes object}

\item{showPlot}{- flag to show plots}
}
\value{
List consisting of:
R_yxz - recruitment by year/sex/size
iN_xmsz - initial population abundance by sex/maturity/shell condition/size
N_yxmsz - population abundance by year/sex/maturity/shell condition/size
F_list - list of fisheries results (see calcNatZ.Fisheries)
N_vyxmsz - survey catches (numbers) by survey/year/sex/maturity/shell condition/size
}
\description{
Function to run the simulation model
}
\details{

}

