% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runSim.BBRKC.R
\name{runSim.BBRKC}
\alias{runSim.BBRKC}
\title{Run the BBRKC simulation model}
\usage{
runSim.BBRKC(showPlot = TRUE)
}
\arguments{
\item{showPlot}{- flag (T/F) to show plots}
}
\description{
Function to run the BBRKC simulation model
}
\details{

}

