% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TextFunctions.R
\name{stripComments}
\alias{stripComments}
\title{Strip comments from elements of a character vector.}
\usage{
stripComments(strv, comment = "#")
}
\arguments{
\item{strv}{- character vector to strip comments from}

\item{comment}{- character indicating remaining text is a comment}
}
\value{
character vector with comments removed.
}
\description{
Function to strip comments from elements of a character vector.
}
\details{


The returned character vector will not have the same number of elements
as the original if the latter had elements that were only comments or 'blank' elements.
Thus, one use of this function could be to remove blank lines, comment-only lines, and
commented text from a character vector created by reading a text file using readLines().
}

