% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TextFunctions.R
\name{stripQuotes}
\alias{stripQuotes}
\title{Strip quotes from a character string vector.}
\usage{
stripQuotes(str, qt = "'")
}
\arguments{
\item{str}{- character vector to strip quotes from}

\item{qt}{- quote character}
}
\value{
character vector w/ quotes stripped out
}
\description{

}

