% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ErrorMessages.R
\name{throwModelTypeError}
\alias{throwModelTypeError}
\title{Throw an unrecognized model type error.}
\usage{
throwModelTypeError(type, str)
}
\arguments{
\item{type}{- model type}

\item{str}{- string to print (name of calling function)}
}
\description{

}
\details{
Stops the R run.
}

