% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeSim.gmacs.R
\name{writeSim.gmacs}
\alias{writeSim.gmacs}
\title{Write simulation results to gmacs input file.}
\usage{
writeSim.gmacs(mc, mp, mr, fn = "gmacs.input.dat", showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration list object}

\item{mp}{- model processes list object}

\item{mr}{- model results list object}

\item{fn}{- output file name}

\item{showPlot}{- flag to show plots}
}
\description{
Function to write simulation results to gmacs input file.
}
\details{

}

