% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcZTM.R
\name{calcZTM}
\alias{calcZTM}
\title{Calculate size transition matrices by year, sex}
\usage{
calcZTM(mc, showPlot = TRUE)
}
\arguments{
\item{mc}{- model configuration object}
}
\value{
list with elements:
prZAM_yxmszz: 6d array with size transition matrices by year/sex/maturity state/shell condition
mnMI_yxz: 3d array with mean size increment at size bin midpoints by year/sex
}
\description{

}

