% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TextFunctions.R
\name{parseNum}
\alias{parseNum}
\title{Parse a number from a character string vector.}
\usage{
parseNum(str)
}
\arguments{
\item{str}{- character vector to parse}
}
\value{
parsed numerical value
}
\description{

}

